+++ src/main/java/com/vehicleManagmentSystem/entity/ApiResidentByRegNo.java
@@ -1,0 +1,19 @@
+package com.vehicleManagmentSystem.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
+
+@Getter
+@Setter
+@Component // use to tell spring load this class this is not entity
+public class ApiResidentByRegNo {
+
+    private String firstName;
+    private String lastName;
+    private String email;
+    private long MobileNo;
+    private String flatNo;
+
+}